Migrate coverage to lcov
Coveralls looks a lot like its dieing:
  - There has been no news update since a promise of a new "news" page
    in Feb 2019
  - cpp-coveralls does not handle modern gcov output, and there is very
    little signs of life on the github page

Its possible at some point we'll need to migrate away from coveralls. As
a first step we now do all the donkey work of lcov generation ourselves.
Currently cpp-coveralls reads that in and loads to coveralls. If
necessary, in the future this can be uploaded to an alternative
provider.